Finalizando o projeto

+++ Checkpoint.Shared/Employee/Dto/GetEmployeeForViewDto.cs
@@ -7,7 +7,7 @@ namespace Checkpoint.Shared.Employee
         public string Name { get; set; }
         public string IdRegistrationCode { get; set; }
         public string Gender { get; set; }
-        public float ValueHourly { get; set; }
+        public decimal ValueHourly { get; set; }
         public decimal Salary { get; set; }
         public bool IsPositionTrust { get; set; }
         public bool IsPjOrClt { get; set; }
